requirements_min.txt
pywebview
PyMuPDF
Pillow
img2pdf
pypdf
